+++ backend/momentreeblog/build.gradle.kts
@@ -58,7 +58,6 @@ dependencies {
 
 	// Oauth2
 	implementation("org.springframework.boot:spring-boot-starter-oauth2-client")
-=======
 }
 
 tasks.withType<Test> {
