--- backend/momentreeblog/src/main/java/com/likelion/momentreeblog/domain/blog/category/controller/CategoryController.java
@@ -1,52 +1,0 @@
-package com.likelion.momentreeblog.domain.blog.category.controller;
-
-import com.likelion.momentreeblog.domain.blog.category.dto.CategoryCreateRequestDto;
-import com.likelion.momentreeblog.domain.blog.category.dto.CategoryUpdateRequestDto;
-import com.likelion.momentreeblog.domain.blog.category.dto.CategoryResponseDto;
-import com.likelion.momentreeblog.domain.blog.category.service.CategoryService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
-
-@RestController
-@RequestMapping("/api/v1/categories")
-@RequiredArgsConstructor
-public class CategoryController {
-
-    private final CategoryService categoryService;
-
-    // 카테고리 생성
-    @PostMapping
-    public ResponseEntity<CategoryResponseDto> createCategory(@RequestBody CategoryCreateRequestDto requestDto) {
-        return ResponseEntity.ok(categoryService.createCategory(requestDto));
-    }
-
-    // 카테고리 단건 조회
-    @GetMapping("/{id}")
-    public ResponseEntity<CategoryResponseDto> getCategory(@PathVariable Long id) {
-        return ResponseEntity.ok(categoryService.getCategory(id));
-    }
-
-    // 모든 카테고리 조회 (선택)
-    @GetMapping
-    public ResponseEntity<List<CategoryResponseDto>> getAllCategories() {
-        return ResponseEntity.ok(categoryService.getAllCategories());
-    }
-
-    // 카테고리 수정
-    @PutMapping("/{id}")
-    public ResponseEntity<CategoryResponseDto> updateCategory(
-            @PathVariable Long id,
-            @RequestBody CategoryUpdateRequestDto requestDto) {
-        return ResponseEntity.ok(categoryService.updateCategory(id, requestDto));
-    }
-
-    // 카테고리 삭제
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteCategory(@PathVariable Long id) {
-        categoryService.deleteCategory(id);
-        return ResponseEntity.noContent().build();
-    }
-}
