--- backend/momentreeblog/src/main/java/com/likelion/momentreeblog/domain/blog/category/dto/CategoryCreateRequestDto.java
@@ -1,11 +1,0 @@
-package com.likelion.momentreeblog.domain.blog.category.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CategoryCreateRequestDto {
-    private String name;
-    private Long blogId;
-}
