--- backend/momentreeblog/src/main/java/com/likelion/momentreeblog/domain/blog/category/dto/CategoryResponseDto.java
@@ -1,20 +1,0 @@
-package com.likelion.momentreeblog.domain.blog.category.dto;
-
-import com.likelion.momentreeblog.domain.blog.category.entity.Category;
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class CategoryResponseDto {
-    private Long id;
-    private String name;
-
-    public static CategoryResponseDto from(Category category) {
-        return CategoryResponseDto.builder()
-                .id(category.getId())
-                .name(category.getName())
-                .build();
-    }
-}
-
