--- backend/momentreeblog/src/main/java/com/likelion/momentreeblog/domain/blog/category/dto/CategoryUpdateRequestDto.java
@@ -1,10 +1,0 @@
-package com.likelion.momentreeblog.domain.blog.category.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class CategoryUpdateRequestDto {
-    private String name;
-}
