--- backend/momentreeblog/src/main/java/com/likelion/momentreeblog/domain/blog/category/entity/Category.java
@@ -1,27 +1,0 @@
-package com.likelion.momentreeblog.domain.blog.category.entity;
-
-import com.likelion.momentreeblog.domain.blog.blog.entity.Blog;
-import com.likelion.momentreeblog.global.jpa.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-import lombok.experimental.SuperBuilder;
-
-@Entity(name = "blogCategory") // 내부 JPA 이름 충돌 방지용
-@Table(name = "categories")
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@SuperBuilder
-public class Category extends BaseEntity {
-
-    @Column(nullable = false)
-    private String name;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "blog_id", nullable = false)
-    private Blog blog;
-
-    public void update(String name) {
-        this.name = name;
-    }
-}
