--- backend/momentreeblog/src/main/java/com/likelion/momentreeblog/domain/blog/category/repository/CategoryRepository.java
@@ -1,7 +1,0 @@
-package com.likelion.momentreeblog.domain.blog.category.repository;
-
-import com.likelion.momentreeblog.domain.blog.category.entity.Category;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-}
