+++ backend/momentreeblog/src/main/java/com/likelion/momentreeblog/domain/board/category/dto/CategoryResponseDto.java
@@ -17,3 +17,4 @@ public class CategoryResponseDto {
                 .build();
     }
 }
+
