+++ backend/momentreeblog/src/main/java/com/likelion/momentreeblog/domain/board/category/entity/Category.java
@@ -1,30 +1,27 @@
 package com.likelion.momentreeblog.domain.board.category.entity;
 
-
-import com.likelion.momentreeblog.domain.board.board.entity.Board;
+import com.likelion.momentreeblog.domain.blog.blog.entity.Blog;
 import com.likelion.momentreeblog.global.jpa.BaseEntity;
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
+import lombok.*;
 import lombok.experimental.SuperBuilder;
 
-import java.util.List;
-
-@Entity
+@Entity(name = "blogCategory") // 내부 JPA 이름 충돌 방지용
 @Table(name = "categories")
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @SuperBuilder
-@ToString
 public class Category extends BaseEntity {
 
     @Column(nullable = false)
     private String name;
 
-    @OneToMany(mappedBy = "category",orphanRemoval = true)
-    private List<Board> boards;
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "blog_id", nullable = false)
+    private Blog blog;
 
+    public void update(String name) {
+        this.name = name;
+    }
 }
