+++ backend/momentreeblog/src/main/java/com/likelion/momentreeblog/domain/user/user/repository/UserRepository.java
@@ -7,4 +7,5 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
+}
